Repeater: Slight change of panel repeater and name change
